pyproject.toml
[tool.poetry]
name = "mirrorme"
version = "0.1.0"
description = ""
authors = ["Manos Ragiadakos"]
license = "BSD 3-Clause License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pillow = "^10.3.0"
opencv-python = "^4.10.0.82"


aiortc = "^1.9.0"
aiohttp = "^3.9.5"
jinja2 = "^3.1.4"
aiohttp-jinja2 = "^1.6"
segno = "^1.6.1"
psutil = "^5.9.8"
[tool.poetry.group.dev.dependencies]
types-psutil = "^5.9.5.20240516"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
mirrorme = 'mirrorme.main:main'

[tool.mypy]
disallow_untyped_calls = false
disallow_untyped_decorators = true
ignore_missing_imports = true
mypy_path = "stubs"
strict_optional = false
warn_redundant_casts = true
warn_unused_ignores = true
